backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
drf-yasg==1.21.8
requests==2.32.3
sqlparse==0.5.5
python-dotenv==1.0.1
Pillow==12.1.1
